refactoring agent base
