+++ reinforcement/widgets/agents/agent_play_mixin.py
@@ -1,5 +1,7 @@
 from functools import partial
 
+from time import sleep
+
 import gym
 
 from Orange.widgets.utils.concurrent import ThreadExecutor
@@ -20,3 +22,33 @@ class AgentPlayMixin():
 
     def stop(self):
         self.playing = False
+
+    def play_action(self, state):
+        pass
+
+    def play_task(self):
+        state = self.enviroment.reset()
+
+        self.enviroment.render()
+
+        self.playing = True
+
+        while self.playing:
+            # pylint: disable=assignment-from-no-return
+            action = self.play_action(state)
+
+            if self.episodes_interval > 0.0:
+                sleep(self.episodes_interval)
+
+            _new_state, _reward, done, _info = self.enviroment.step(action)
+
+            self.enviroment.render()
+
+            if done:
+                if self.games_interval > 0.0:
+                    sleep(self.games_interval)
+
+                self.enviroment.reset()
+                self.enviroment.render()
+
+        self.enviroment.close()
